pyproject.toml
[project]
name = "onetool-mcp"
version = "1.0.0"
description = "MCP server exposing a single 'run' tool for LLM code execution"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0" }
authors = [{ name = "Gavin Las", email = "beycom99@gmail.com" }]
keywords = ["mcp", "llm", "code-execution", "tools", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "fastmcp>=2.14.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mcp>=1.25.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.1",
    # Browser inspector (requires: playwright install chromium)
    "playwright>=1.57.0",
    # Browser CLI (onetool-browse)
    "questionary>=2.1.1",
    "pillow>=12.1.0",
    # Snippets template engine
    "jinja2>=3.1.6",
    # Gemini grounding search
    "google-genai>=1.59.0",
    # Message tool async file I/O
    "aiofiles>=25.1.0",
    # Docstring parsing for ot.help() and ot.tools()
    "docstring-parser>=0.17",
]

[project.urls]
Homepage = "https://github.com/beycom/onetool"
Repository = "https://github.com/beycom/onetool"
Documentation = "https://github.com/beycom/onetool/tree/main/docs"
Issues = "https://github.com/beycom/onetool/issues"

[project.optional-dependencies]
# File tool safe deletion (moves to trash instead of permanent delete)
file = ["send2trash>=2.1.0"]

[project.scripts]
ot-serve = "ot_serve.cli:cli"
ot-bench = "ot_bench.cli:cli"
ot-browse = "ot_browse.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ot", "src/ot_tools", "src/ot_sdk", "src/ot_serve", "src/ot_bench", "src/ot_browse"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-sugar>=1.1.1",
    "ruff>=0.14.13",
    "types-pyyaml>=6.0.12.20250915",
    # Tool dependencies for testing (these use inline script metadata in production)
    "duckdb>=1.0.0",      # for ot_tools.code_search
    "openpyxl>=3.1.5",    # for ot_tools.excel
    "send2trash>=2.1.0",  # for ot_tools.file safe deletion
    "sqlalchemy>=2.0.45", # for ot_tools.db
    "trafilatura>=2.0.0", # for ot_tools.web_fetch
]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]
cache-dir = "tmp/.ruff_cache"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (required by typer)
]

[tool.ruff.lint.isort]
known-first-party = ["ot", "ot_tools", "ot_sdk", "ot_serve", "ot_bench", "ot_browse"]

[tool.ruff.lint.per-file-ignores]
# CLI files need Path and CaseResult at runtime for typer type annotations and callbacks
"src/ot_serve/*.py" = ["TC001", "TCH003"]
"src/ot_bench/*.py" = ["TC001", "TCH003"]
# Browser CLI uses Path and Annotation at runtime, SIM105 for event handlers
"src/ot_browse/*.py" = ["TC001", "TC003", "SIM105"]
# Tools use namespace declarations before imports for tool discovery
"src/ot_tools/*.py" = ["E402"]
# Executor modules use Path/Callable at runtime for tool loading and worker management
"src/ot/executor/*.py" = ["TC003"]
# SDK modules use Callable/Generator at runtime for worker functions
"src/ot_sdk/*.py" = ["TC003"]

[tool.mypy]
python_version = "3.11"
cache_dir = "tmp/.mypy_cache"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src/ot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src/ot_tools"]
addopts = "-v"
asyncio_mode = "auto"
cache_dir = "tmp/.pytest_cache"
markers = [
    # Speed tiers (required - pick one)
    "smoke: Quick sanity checks (<1s each)",
    "unit: Fast isolated tests (<1s each, no I/O)",
    "integration: End-to-end tests (may be slow)",
    "slow: Long-running tests (>10s)",

    # Component tags (required - pick one)
    "serve: MCP server tests",
    "bench: Benchmark harness tests",
    "browse: Browser inspector tests",
    "core: Core library tests",
    "tools: Tool implementation tests",
    "pkg: Package tests",
    "spec: Specification tests",

    # Dependency tags (add as needed)
    "network: Requires network access",
    "api: Requires API keys",
    "playwright: Requires playwright browsers",
    "docker: Requires Docker daemon",
]

[tool.deptry]
# Ignore internal packages (DEP001) - these are first-party, not external deps
per_rule_ignores = { DEP001 = ["ot", "ot_tools", "ot_sdk", "ot_bench", "ot_browse", "ot_serve", "anyio"] }
# Map package names where import name differs
[tool.deptry.package_module_name_map]
pyyaml = "yaml"
pillow = "PIL"
pydantic-settings = "pydantic_settings"
google-genai = "google"

[tool.coverage.html]
directory = "tmp/htmlcov"
